rename google components
