+++ src/app/components/google_map/google_map.tsx
@@ -16,14 +16,7 @@ export default async function GoogleMap({ query }: { query: string }) {
 
   return (
     <Box className={classes.map}>
-      {
-        <GoogleMapsEmbed
-          apiKey={mapApiKey}
-          mode="place"
-          q={query}
-          zoom={"15"}
-        />
-      }
+      <GoogleMapsEmbed apiKey={mapApiKey} mode="place" q={query} zoom={"15"} />
     </Box>
   );
 }
